feat: rosetta data api and construction validation with rosetta-cli
* test: added test for rosetta cli check:construction and check:data

* ci: added github workflow to test  roseta-cli:data

* ci: add github action for rosetta-cli: construction

* refactor: refactor cli tests

* fix: use localhoust for github actions in rosetta config file

* fix: github actions

* fix: rebase issue

* refactor: update rosetta-cli env files path

* refactor: remove hard coded cli container and use rosetta-cli ouput

* perf: remove unnecessary wait of 20 seconds
