+++ .github/workflows/stacks-blockchain-api.yml
@@ -266,6 +266,126 @@ jobs:
         uses: codecov/codecov-action@v1
         if: always()
 
+  test-rosetta-cli-data:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Use Node.js
+        uses: actions/setup-node@v2
+        with:
+          node-version: '14'
+          check-latest: true
+
+      - name: Cache node modules
+        uses: actions/cache@v2
+        env:
+          cache-name: cache-node-modules
+        with:
+          path: |
+            ~/.npm
+            node_modules
+          key: ${{ runner.os }}-build-${{ env.cache-name }}-${{ hashFiles('**/package-lock.json') }}
+          restore-keys: |
+            ${{ runner.os }}-build-${{ env.cache-name }}-
+            ${{ runner.os }}-build-
+            ${{ runner.os }}-
+
+      - name: Install deps
+        run: npm install --audit=false
+
+      - name: Setup env vars
+        run: echo "STACKS_CORE_EVENT_HOST=http://0.0.0.0">> $GITHUB_ENV
+
+      - name: Setup cli enviroment variable
+        run: | 
+          echo STACKS_BLOCKCHAIN_API_HOST=0.0.0.0>> .env
+          echo STACKS_CORE_PROXY_HOST=0.0.0.0 >> .env
+          echo STACKS_CORE_RPC_HOST=0.0.0.0 >> .env
+          echo STACKS_CORE_EVENT_HOST=0.0.0.0 >> .env 
+          echo BTC_RPC_HOST=http://0.0.0.0 >> .env 
+
+      - name: Setup integration environment
+        run: |
+          sudo ufw disable
+          npm run devenv:deploy -- -d
+          npm run devenv:logs -- --no-color &> docker-compose-logs.txt &
+
+      - name: Run tests
+        run: sudo npm run test:rosetta-cli:data
+
+      - name: Print integration environment logs
+        run: cat docker-compose-logs.txt
+        if: failure()
+
+      - name: Teardown integration environment
+        run: npm run devenv:stop
+        if: always()
+
+      - name: Upload coverage to Codecov
+        uses: codecov/codecov-action@v1
+        if: always()
+
+  test-rosetta-cli-construction:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Use Node.js
+        uses: actions/setup-node@v2
+        with:
+          node-version: '14'
+          check-latest: true
+
+      - name: Cache node modules
+        uses: actions/cache@v2
+        env:
+          cache-name: cache-node-modules
+        with:
+          path: |
+            ~/.npm
+            node_modules
+          key: ${{ runner.os }}-build-${{ env.cache-name }}-${{ hashFiles('**/package-lock.json') }}
+          restore-keys: |
+            ${{ runner.os }}-build-${{ env.cache-name }}-
+            ${{ runner.os }}-build-
+            ${{ runner.os }}-
+
+      - name: Install deps
+        run: npm install --audit=false
+
+      - name: Setup env vars
+        run: echo "STACKS_CORE_EVENT_HOST=http://0.0.0.0">> $GITHUB_ENV
+      
+      - name: Setup cli enviroment variable
+        run: | 
+          echo STACKS_BLOCKCHAIN_API_HOST=0.0.0.0>> .env
+          echo STACKS_CORE_PROXY_HOST=0.0.0.0 >> .env
+          echo STACKS_CORE_RPC_HOST=0.0.0.0 >> .env
+          echo STACKS_CORE_EVENT_HOST=0.0.0.0 >> .env 
+          echo BTC_RPC_HOST=http://0.0.0.0 >> .env 
+
+      - name: Setup integration environment
+        run: |
+          sudo ufw disable
+          npm run devenv:deploy -- -d
+          npm run devenv:logs -- --no-color &> docker-compose-logs.txt &
+
+      - name: Run tests
+        run: sudo npm run test:rosetta-cli:construction
+
+      - name: Print integration environment logs
+        run: cat docker-compose-logs.txt
+        if: failure()
+
+      - name: Teardown integration environment
+        run: npm run devenv:stop
+        if: always()
+
+      - name: Upload coverage to Codecov
+        uses: codecov/codecov-action@v1
+        if: always()
+
   test-tokens:
     runs-on: ubuntu-latest
     steps:
