+++ docker-compose.dev.rosetta-cli.yml
@@ -4,4 +4,7 @@ services:
     build:
       context: ./rosetta-cli-config
       dockerfile: docker/Dockerfile
-    command: /bin/rosetta-cli --configuration-file /app/rosetta-config-docker.json check:construction
+    command: ${CMD}
+    volumes:
+      - ${OUTPUT}
+
