+++ rosetta-cli-config/rosetta-config-docker.json
@@ -6,9 +6,7 @@
   "online_url": "http://host.docker.internal:3999/rosetta/v1",
   "data_directory": "",
   "http_timeout": 10,
-  "sync_concurrency": 8,
-  "transaction_concurrency": 16,
-  "tip_delay": 5,
+  "tip_delay": 500,
   "data": {
     "active_reconciliation_concurrency": 16,
     "inactive_reconciliation_concurrency": 4,
@@ -18,80 +16,70 @@
     "log_balance_changes": true,
     "log_reconciliations": false,
     "ignore_reconciliation_error": false,
-    "exempt_accounts": "/app/exempt_accounts.json",
-    "bootstrap_balances": "/app/bootstrap_balances.json",
-    "interesting_accounts": "/app/interesting_accounts.json",
-    "historical_balance_disabled": true,
+    "exempt_accounts": "exempt_accounts.json",
+    "bootstrap_balances": "bootstrap_balances.json",
+    "historical_balance_enabled": true,
     "reconciliation_disabled": false,
     "inactive_discrepency_search_disabled": false,
     "balance_tracking_disabled": false,
     "end_conditions": {
-      "reconciliation_coverage": 0.95,
-      "index": 100
+      "reconciliation_coverage": {
+        "coverage": 0.80,
+        "index": 100
+      }
     },
     "results_output_file": "./rosetta-output/rosetta-cli-output.json"
   },
   "construction": {
     "offline_url": "http://host.docker.internal:3999/rosetta/v1",
-    "currency": {
-      "symbol": "STX",
-      "decimals": 18
-    },
-    "minimum_balance": "0",
-    "maximum_fee": "10000000000000000",
-    "curve_type": "secp256k1",
-    "accounting_model": "account",
-    "scenario": [
+    "max_offline_connections": 0,
+    "stale_depth": 0,
+    "broadcast_limit": 0,
+    "ignore_broadcast_failures": false,
+    "clear_broadcasts": false,
+    "broadcast_behind_tip": false,
+    "block_broadcast_limit": 0,
+    "rebroadcast_all": false,
+    "constructor_dsl_file": "stacks.ros",
+    "prefunded_accounts": [
       {
-        "operation_identifier": {
-          "index": 0
+        "privkey": "21d43d2ae0da1d9d04cfcaac7d397a33733881081f0b2cd038062cf0ccbb7526",
+        "account_identifier": {
+          "address": "ST11NJTTKGVT6D1HY4NJRVQWMQM7TVAR091EJ8P2Y"
         },
-        "type": "NATIVE_TRANSFER",
-        "status": "",
-        "account": {
-          "address": "{{ SENDER }}"
-        },
-        "amount": {
-          "value": "{{ SENDER_VALUE }}",
-          "currency": {
-            "symbol": "STX",
-            "decimals": 6
-          }
+        "curve_type": "secp256k1",
+        "currency": {
+          "symbol": "STX",
+          "decimals": 6
         }
       },
       {
-        "operation_identifier": {
-          "index": 1
+        "privkey": "c71700b07d520a8c9731e4d0f095aa6efb91e16e25fb27ce2b72e7b698f8127a",
+        "account_identifier": {
+          "address": "ST1HB1T8WRNBYB0Y3T7WXZS38NKKPTBR3EG9EPJKR"
         },
-        "related_operations": [
-          {
-            "index": 0
-          }
-        ],
-        "type": "NATIVE_TRANSFER",
-        "status": "",
-        "account": {
-          "address": "{{ RECIPIENT }}"
+        "curve_type": "secp256k1",
+        "currency": {
+          "symbol": "STX",
+          "decimals": 6
+        }
+      },
+      {
+        "privkey": "e75dcb66f84287eaf347955e94fa04337298dbd95aa0dbb985771104ef1913db",
+        "account_identifier": {
+          "address": "STRYYQQ9M8KAF4NS7WNZQYY59X93XEKR31JP64CP"
         },
-        "amount": {
-          "value": "{{ RECIPIENT_VALUE }}",
-          "currency": {
-            "symbol": "STX",
-            "decimals": 6
-          }
+        "curve_type": "secp256k1",
+        "currency": {
+          "symbol": "STX",
+          "decimals": 6
         }
       }
     ],
-    "confirmation_depth": 5,
-    "stale_depth": 30,
-    "broadcast_limit": 3,
-    "ignore_broadcast_failures": false,
-    "change_scenario": null,
-    "clear_broadcasts": false,
-    "broadcast_behind_tip": false,
-    "block_broadcast_limit": 5,
-    "rebroadcast_all": false,
-    "new_account_probability": 0.5,
-    "max_addresses": 200
+    "end_conditions": {
+      "create_account": 10,
+      "transfer": 10
+    },
+    "results_output_file": "./rosetta-output/rosetta-cli-output-const.json"
   }
 }
