+++ src/tests-rosetta-cli-construction/envs/env.construction
@@ -1,0 +1,2 @@
+CMD=/bin/rosetta-cli --configuration-file /app/rosetta-config-docker.json check:construction
+OUTPUT=./rosetta-output-construction:/app/rosetta-output
