+++ src/tests-rosetta-cli-data/contracts/hello-world.clar
@@ -1,0 +1,5 @@
+(define-public (say-hi)
+   (ok "hello world"))
+
+(define-public (echo-number (val int))
+   (ok val))
