+++ src/tests-rosetta-cli-data/envs/env.data
@@ -1,0 +1,2 @@
+CMD=/bin/rosetta-cli --configuration-file /app/rosetta-config-docker.json check:data
+OUTPUT=./rosetta-output:/app/rosetta-output
